MNT: Fix mybinder.
The combination of mybinder, anaconda, and conda-forge results in the
invalid downgrade of jpeg from 9 to 8, causing a problem with the
conda-forge packages (this is caused by a bug in the old version of
anaconda on the mybinder docker image). Use our our Docker image now to
work around this.